requirements.txt
# Anuj Bot - Working Requirements
# Essential packages for bot functionality

# Telegram Bot Framework
python-telegram-bot==20.7

# Database - MongoDB
pymongo==4.6.0
motor==3.3.2

# Environment Variables
python-dotenv==1.0.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber

# Image Processing
Pillow==10.1.0
opencv-python

# AI and NLP
openai==1.3.8
nltk
textblob

# Web Scraping and HTTP
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2

# Data Processing
numpy
pandas

# Text Processing
python-docx==1.1.0
markdown

# Async Support
aiofiles

# Logging and Monitoring
loguru

# Utilities
python-dateutil
pytz
schedule

# OCR for Image Text Extraction
pytesseract

# Math and Scientific Computing (Optional - for advanced features)
# sympy
# matplotlib

# Advanced OCR (Optional - requires PyTorch, large download)
# easyocr

# Security
cryptography

